Sanjana

+++ ToDoList/app/src/main/java/com/example/to_dolist/ui/fragments/DetailsFragment.kt
@@ -1,11 +1,14 @@
 package com.example.to_dolist.ui.fragments
 
 import android.annotation.SuppressLint
+import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.appcompat.widget.PopupMenu
 import androidx.core.os.bundleOf
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
@@ -19,6 +22,9 @@ import com.google.gson.reflect.TypeToken
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
 import java.text.SimpleDateFormat
+import android.content.Intent
+import androidx.appcompat.app.AlertDialog
+import android.content.DialogInterface
 
 @AndroidEntryPoint
 class DetailsFragment : Fragment() {
@@ -28,7 +34,7 @@ class DetailsFragment : Fragment() {
     private val viewModel: TodoListViewModel by viewModels()
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        savedInstanceState: Bundle?,
     ): View? {
         binding = FragmentDetailsBinding.inflate(inflater, container, false)
         return binding.root
@@ -45,19 +51,46 @@ class DetailsFragment : Fragment() {
         val type = object : TypeToken<TodoListItem>() {}.type
         todo = Gson().fromJson(data, type)
         binding.showTaskTitle.setText(todo.title)
-        val datetime = SimpleDateFormat("EE dd MMM yyyy hh:mm a").format(todo.time)
+        val datetime = SimpleDateFormat("EE dd MMM yyyy hh:mm aa").format(todo.time)
         binding.showTaskDateTime.setText(datetime)
         binding.showSubTask.setText(todo.todo.joinToString("\n"))
-
     }
+
     private fun initOnClick() {
         binding.apply {
-            deleteBtn.setOnClickListener {
-             viewModel.deleteTodoList(todo)
-            }
-            editBtn.setOnClickListener {
-                val dataFromDetailsToUpdate = bundleOf(Constants.DATA_FROM_DETAILS_TO_UPDATE to Gson().toJson(todo))
-                findNavController().navigate(R.id.action_detailsFragment_to_updateFragment,dataFromDetailsToUpdate)
+            moreVertical.setOnClickListener {
+                val menuItemView =
+                    view!!.findViewById<View>(com.example.to_dolist.R.id.moreVertical)
+                val popupMenu = PopupMenu(activity!!, menuItemView)
+                popupMenu.menuInflater.inflate(com.example.to_dolist.R.menu.popup_menu,
+                    popupMenu.menu)
+
+                popupMenu.setOnMenuItemClickListener { item ->
+                    when (item.itemId) {
+                        R.id.deleteBtn -> {
+                            AlertDialog.Builder(activity!!)
+                                .setTitle("Confirm delete")
+                                .setIcon(R.drawable.ic_warning)
+                                .setMessage("Are you sure you want to delete?")
+                                .setPositiveButton("Yes") { dialog, which ->
+                                    viewModel.deleteTodoList(todo)
+                                    Toast.makeText(activity,
+                                        "Successfully deleted",
+                                        Toast.LENGTH_SHORT).show();
+                                    findNavController().navigate(R.id.action_detailsFragment_to_homeFragment)
+                                }
+                                .setNegativeButton("No", null).show()
+                        }
+                        R.id.editBtn -> {
+                            val dataFromDetailsToUpdate =
+                                bundleOf(Constants.DATA_FROM_DETAILS_TO_UPDATE to Gson().toJson(todo))
+                            findNavController().navigate(R.id.action_detailsFragment_to_updateFragment,
+                                dataFromDetailsToUpdate)
+                        }
+                    }
+                    true
+                }
+                popupMenu.show()
             }
         }
     }
